% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumSensCorrect.R
\name{HumSensCorrect}
\alias{HumSensCorrect}
\title{Function to clean relative humidity data from self-built sensors}
\usage{
HumSensCorrect(v, t.inter = 5.404, t.slope = 1.0745, min.rescale = 7.5,
  max.rescale = 98.8, poly.a = 0.002462, poly.b = 0.732, poly.c = 7.579)
}
\arguments{
\item{v}{Vector that contains the measured relative humidity values.}

\item{t.inter}{Intercept of the temperature correction. Default provided. Numeric.}

\item{t.slope}{Slope of the temperature correction. Default provided. Numeric.}

\item{min.rescale}{Minimal relative humidity that was found at the site. Used set the range of the relative humidity values. Default provided from the calibration experiment. Numeric.}

\item{max.rescale}{Maximum relative humidity that was found at the site. Used set the range of the relative humidity values. Default provided from the calibration experiment. Numeric.}

\item{poly.a}{First coefficient of the calibration polynom (a * v^2 + b * v + c). Default provided from calibration experiment.}

\item{poly.b}{First coefficient of the calibration polynom (a * v^2 + b * v + c). Default provided from calibration experiment.}

\item{poly.c}{First coefficient of the calibration polynom (a * v^2 + b * v + c). Default provided from calibration experiment.}
}
\value{
Vector with the calibrated and re-scaled relative humidity values.
}
\description{
Function to clean relative humidity data from self-built sensors
}
\examples{
\dontrun{
co.hum$Hum.func <- HumSensCorrect(co.hum$Hum_Avg)
}
}
