% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempSensCorrect.R
\name{TempSensCorrect}
\alias{TempSensCorrect}
\title{Function to correct air-temperature data from self-built sensors}
\usage{
TempSensCorrect(v, glob.rad, inter = -2.899, slope = 1.181)
}
\arguments{
\item{v}{Vector with temperature data from self-built air-temperature sensor that needs correction.}

\item{glob.rad}{Vector that contains values of global radiation (W/m2) for each value of v. i.e. must be same length as v.}

\item{inter}{intercept of relationship between self-built sensor and calibrated sensor. Default determined through a calibration experiment.}

\item{slope}{slope of relationship between self-built sensor and calibrated sensor. Default determined through a calibration experiment.}
}
\value{
Vector with the calibrated values.
}
\description{
Function to correct air-temperature data from self-built sensors
}
\examples{
\dontrun{
Temp.func <- TempSensCorrect(co.temp$Temp_Avg, glob.rad = co.temp$Ave.GSR..W.m.2.)
}
}
